3_Neural Network from scratch/DeepPip/requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0
PyYAML>=6.0